rag_frontend/requirements.txt
# rag_frontend/requirements.txt
streamlit
chromadb
mysql-connector-python
pysqlite3-binary
google-generativeai
